# FBXLoader reduce casting between types and refactor typecasting