A couple of changes here:

* Combined `parseFloatArray` and `parseIntArray` into `parseNumberArray` (no need to worry about the type as parseFloat doesn't accidentally change `ints` to `floats`

* removed `parseMatrixArray` (it's was being called in only one place after this refactoring, and was a single line function so I inlined it). 

* where possible, make the `TextParser` do type checking internally (ideally, `parseNumberArray` should be called only inside the `TextParser` but there are a couple of cases I haven't found yet, so marked those as TODO). 

* `BinaryParser` was casting arrays of numbers into strings, which were then passed to `parseFloatArray` or`parseIntArray` to be converted back into numbers. This was to make the return the same as the TextParser. To remove this I added a check so that if `parseNumberArray` is passed an Array rather than String it will just return it.  Once the final types are caught in TextParser this check won't be needed (both parsers should return arrays ).

The final change is the big one here - I would expect Binary format parsing to be significantly faster after this, since the arrays being cast to strings and then back to arrays were the ones holding data such as vertex positions, uvs, normals, keyframe data and so on.