@looeee 2017-10-24
Testing on a couple of binary files, here are the speedups in parsing time:

mixamo model with mocap animations: 1977ms-> 1003ms (49% faster)
nao robot model with animations: 3367ms -> 1916ms (44% faster)
static scene with lots of embedded textures: 4363ms -> 1995ms (55% faster)

And of course
xsi_man_skinning : 2090ms -> 1589ms (23% faster) - as expected not quite such as speedup since the file is not actually compressed (doesn't require inflate.js)

😁